+++ ci-scripts/deploy/deploy.sh
@@ -36,12 +36,20 @@ if [[ -n ${PINGONE} ]]; then
   python3 ${PROJECT_DIR}/ci-scripts/deploy/ping-one/p1_env_setup_and_teardown.py Setup
 fi
 
-DEPLOY_FILE=/tmp/deploy.yaml
+deploy_file=/tmp/deploy.yaml
 
-build_dev_deploy_file "${DEPLOY_FILE}"
+# First, we need to deploy cert-manager. This is due to it using Dynamic Admission Control - Mutating Webhooks which
+# must be available before we make use cert-manager
+kubectl apply -f "${PROJECT_DIR}/k8s-configs/cluster-tools/base/cert-manager/base/cert-manager.yaml"
 
-kubectl apply -f "${DEPLOY_FILE}"
+# Build file while cert-manager webhook service coming up to save time
+build_dev_deploy_file "${deploy_file}"
 
-check_if_ready ${NAMESPACE}
+# Wait until the webhook deployment is fully available
+wait_for_rollout "deployment/cert-manager-webhook" "cert-manager" "20"
+
+kubectl apply -f "${deploy_file}"
+
+check_if_ready "${NAMESPACE}"
 
 popd  > /dev/null 2>&1